api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
pymongo==4.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
gunicorn==21.2.0